% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_shrinkage.R
\name{rev_npn_shrinkage}
\alias{rev_npn_shrinkage}
\title{Reverse NPN (shrunken ECDF) transformation}
\usage{
rev_npn_shrinkage(mat_trans, parameters = NULL)
}
\arguments{
\item{mat_trans}{Transformed matrix}

\item{parameters}{Parameters of the transformation}
}
\value{
The matrix before transformation
}
\description{
Compute the reverse of the nonparanormal transformation (with shrunken ECDF)
Note that, if the matrix to be reverse transformed has values outside of the reference parameters, they will be clipped.
}
\examples{

# Reverse a previous transformation
my_mat <- matrix(rnorm(6, 10, 1), ncol = 2)
transformed_mat <- transform_npn_shrinkage(my_mat)
rev_npn_shrinkage(transformed_mat$mat, transformed_mat$parameters)

# Given a matrix in NPN scale, project it back onto original scale
y <- matrix(c(-0.5, 0.2), ncol =1)
rev_npn_shrinkage(y, transformed_mat$parameters)
}
