% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_truncation.R
\name{rev_npn_truncation}
\alias{rev_npn_truncation}
\title{Reverse NPN (truncation) transformation}
\usage{
rev_npn_truncation(mat_trans, parameters = NULL)
}
\arguments{
\item{mat_trans}{Transformed matrix}

\item{parameters}{Parameters of the transformation}
}
\value{
The matrix before transformation
}
\description{
Compute the reverse of the nonparanormal transformation (with truncation)
Note that, if the matrix to be reverse transformed has close to the limits of the reference parameters, they will be clipped.
}
\examples{
# Reverse a previous transformation
my_mat <- matrix(rnorm(6, 10, 1), ncol = 2)
transformed_mat <- transform_npn_shrinkage(my_mat)
rev_npn_shrinkage(transformed_mat$mat, transformed_mat$parameters)

# Given a matrix in NPN scale, project it back onto original scale
y <- matrix(c(-1, 1), ncol = 1)
rev_npn_shrinkage(y, transformed_mat$parameters)
}
