% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score.R
\name{reverse_transform_zscore}
\alias{reverse_transform_zscore}
\title{Reverse Z-score transformation}
\usage{
reverse_transform_zscore(mat_trans, parameters = NULL)
}
\arguments{
\item{mat_trans}{Transformed matrix}

\item{parameters}{The transformation parameters}
}
\value{
The original matrix (untransformed)
}
\description{
Reverse Z-score transformation
}
