% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score.R
\name{transform_zscore}
\alias{transform_zscore}
\title{Z score transformation also returning the transformation parameters}
\usage{
transform_zscore(mat, parameters = NULL)
}
\arguments{
\item{mat}{Matrix to transform (transforming each column)}

\item{parameters}{If not \code{NULL}, project onto the scale defined by those parameters}
}
\value{
A list with two components: the transformed matrix and the parameters
}
\description{
Z score transformation also returning the transformation parameters
}
